requirements.txt
typer>=0.15.0
rich>=13.7.0
pillow>=10.2.0
pandas>=2.2.0
